Further C++14 compatibility changes for netservices2
Addresses additional errors found after initial C++14 refactor:
- Fixed signed/unsigned warning in HttpBuffer.cpp.
- Corrected AddField call type mismatch in HttpRequest.cpp.
- Refactored HttpFields.h/cpp and HttpSerializer.h/cpp to remove
  std::string_view, std::optional, and std::byte, aligning with
  C++14 patterns.

KNOWN ISSUE: HttpParser.cpp still has critical syntax errors
(likely an unclosed scope in ParseFields or similar) that prevent
its compilation. These errors cause cascading parsing failures for
subsequent definitions in that file. Further debugging is required
for HttpParser.cpp.

+++ src/kits/network/libnetservices2/HttpBuffer.cpp
@@ -198,7 +198,7 @@ HttpBuffer::Data(size_t& length) const noexcept
 HttpBuffer&
 HttpBuffer::operator<<(const BString& data)
 {
-	if (data.Length() > (fBuffer.capacity() - fBuffer.size())) { // Changed .length() to .Length()
+	if (static_cast<size_t>(data.Length()) > (fBuffer.capacity() - fBuffer.size())) {
 		throw BNetworkRequestError(__PRETTY_FUNCTION__, BNetworkRequestError::ProtocolError,
 			"No capacity left in buffer to append data.");
 	}
